+++ src/app/sender.html
@@ -8,6 +8,6 @@
   </head>
   <body>
     <div id="root"></div>
-    <script type="module" src="/sender.tsx"></script>
+    <script type="module" src="/sender/sender.tsx"></script>
   </body>
 </html>
