+++ src/app/sender/Streaming.tsx
@@ -1,112 +1,21 @@
-import './globals.css';
-import type { FormEvent } from 'react';
-import { StrictMode, useEffect, useState } from 'react';
+import Button from '../components/Button/Button';
+import { usePersistentState } from '../utils/storage';
+import { patchLiveShareToken } from '../components/ShareLiveStatus/api';
+import { copyTextToClipboard } from '../utils/clipboard';
+import useShareLivePosition from '../utils/useShareLivePosition';
+import { writeError } from '../utils/logs';
+import CopyIcon from '../components/icons/CopyIcon';
+import RefreshIcon from '../components/icons/RefreshIcon';
+import BroadcastIcon from '../components/icons/BroadcastIcon';
+import { classNames } from '../utils/styles';
+import { getGameInfo, useIsNewWorldRunning } from '../utils/games';
+import { toast } from 'react-toastify';
 import { v4 as uuid } from 'uuid';
-import ReactDOM from 'react-dom';
-import AppHeader from './components/AppHeader/AppHeader';
-import type { Position } from './contexts/PositionContext';
-import { SettingsProvider } from './contexts/SettingsContext';
-import type { AccountDTO } from './contexts/UserContext';
-import { useAccount, UserProvider } from './contexts/UserContext';
-import { waitForOverwolf } from './utils/overwolf';
-import styles from './Sender.module.css';
-import { getGameInfo, useIsNewWorldRunning } from './utils/games';
-import Ads from './components/Ads/Ads';
-import steamSrc from './components/User/steam.png';
-import { fetchJSON } from './utils/api';
-import { ToastContainer, toast } from 'react-toastify';
-import 'react-toastify/dist/ReactToastify.css';
-import Button from './components/Button/Button';
-import { setJSONItem, usePersistentState } from './utils/storage';
-import { patchLiveShareToken } from './components/ShareLiveStatus/api';
-import { copyTextToClipboard } from './utils/clipboard';
-import useShareLivePosition from './utils/useShareLivePosition';
-import { writeError } from './utils/logs';
-import CopyIcon from './components/icons/CopyIcon';
-import RefreshIcon from './components/icons/RefreshIcon';
-import BroadcastIcon from './components/icons/BroadcastIcon';
-import { classNames } from './utils/styles';
-
-const { VITE_API_ENDPOINT = '' } = import.meta.env;
-
-function Welcome(): JSX.Element {
-  const [verifyingSessionId, setVerifyingSessionId] = useState('');
-  const { setAccount } = useAccount();
-
-  useEffect(() => {
-    if (!verifyingSessionId) {
-      return;
-    }
-    const toastId = toast('Waiting for Steam Sign In', {
-      type: 'info',
-      autoClose: false,
-    });
-    const intervalId = setInterval(async () => {
-      try {
-        const init: RequestInit = {};
-        if (verifyingSessionId) {
-          init.headers = {
-            'x-session-id': verifyingSessionId,
-            'x-prevent-logout': 'true',
-          };
-        }
-        const newAccount = await fetchJSON<AccountDTO>(
-          `/api/auth/account`,
-          init
-        );
-        setAccount(newAccount);
-        setVerifyingSessionId('');
-      } catch (error) {
-        // Keep waiting
-      }
-    }, 3000);
-    return () => {
-      clearInterval(intervalId);
-      toast.dismiss(toastId);
-    };
-  }, [verifyingSessionId]);
-
-  const handleLogin = async () => {
-    const newSessionId = await fetchJSON<string>('/api/auth/session');
-
-    const url = `${VITE_API_ENDPOINT}/api/auth/steam?sessionId=${newSessionId}`;
-    overwolf.utils.openUrlInDefaultBrowser(url);
-
-    setVerifyingSessionId(newSessionId);
-  };
-
-  return (
-    <div className={styles.welcome}>
-      <p>
-        Connect to{' '}
-        <a href="https://aeternum-map.gg" target="_blank">
-          https://aeternum-map.gg
-        </a>
-        , an interactive New World map with routes and community managed
-        markers.
-      </p>
-      <ul>
-        <li>🚀 Live Tracking of your In-Game position</li>
-        <li>🤗 See your friends by using the same token</li>
-        <li>🔀 Farming/Marker Routes</li>
-        <li>✅ Check markers as done (like lore documents)</li>
-        <li>🗺️ Minimap view</li>
-        <li>
-          🤷‍♂️{' '}
-          <a
-            href="https://discord.com/channels/320539672663031818/896014490808745994/911185526210576394"
-            target="_blank"
-          >
-            Conforms to AGS ToS
-          </a>
-        </li>
-      </ul>
-      <button onClick={handleLogin}>
-        <img src={steamSrc} alt="Sign in through Steam" />
-      </button>
-    </div>
-  );
-}
+import type { FormEvent } from 'react';
+import type { Position } from '../contexts/PositionContext';
+import { useEffect, useState } from 'react';
+import { useAccount } from '../contexts/UserContext';
+import styles from './Streaming.module.css';
 
 function Streaming(): JSX.Element {
   const { account, logoutAccount } = useAccount();
@@ -317,36 +226,4 @@ function Streaming(): JSX.Element {
   );
 }
 
-function Sender(): JSX.Element {
-  const { account } = useAccount();
-
-  return (
-    <div className={styles.container}>
-      <AppHeader />
-      <main className={styles.main}>
-        {account ? <Streaming /> : <Welcome />}
-      </main>
-      <Ads active />
-      <ToastContainer
-        theme="dark"
-        pauseOnHover={false}
-        autoClose={3000}
-        pauseOnFocusLoss={false}
-        style={{ marginTop: 32 }}
-      />
-    </div>
-  );
-}
-
-waitForOverwolf().then(() => {
-  ReactDOM.render(
-    <StrictMode>
-      <SettingsProvider>
-        <UserProvider>
-          <Sender />
-        </UserProvider>
-      </SettingsProvider>
-    </StrictMode>,
-    document.querySelector('#root')
-  );
-});
+export default Streaming;
