+++ src/app/sender/Welcome.module.css
@@ -1,0 +1,5 @@
+.welcome {
+  display: grid;
+  align-content: center;
+  justify-content: center;
+}
